+++ ros2/nodes/ti_vision_cnn/launch/semseg_cnn_launch.py
@@ -69,7 +69,7 @@ def generate_launch_description():
     # DL model path
     dl_model_path = DeclareLaunchArgument(
         name="dl_model_path",
-        default_value=TextSubstitution(text="/opt/model_zoo/TVM-SS-5818-deeplabv3lite-mobv2-qat-robokit-768x432"),
+        default_value=TextSubstitution(text="/opt/model_zoo/ONR-SS-8818-deeplabv3lite-mobv2-qat-robokit-768x432"),
         # default_value=TextSubstitution(text="/opt/model_zoo/ONR-SS-8818-deeplabv3lite-mobv2-qat-robokit-768x432"),
         description='DL model path'
     )